app/python/requirements.txt
tensorflow>=2.13.0
mediapipe==0.10.14
opencv-python>=4.8.0
numpy>=1.24.0
pyserial>=3.5
flask>=2.3.0
flask-cors>=4.0.0
tensorflow>=2.13.0
mediapipe>=0.10.0
opencv-python>=4.8.0
numpy>=1.24.0
pyserial>=3.5
flask>=2.3.0
flask-cors>=4.0.0
